LayoutTests/canvas/philip/tests/2d.shadow.stroke.join.3-expected.txt

< [index] >
2d.shadow.stroke.join.3
Shadows are drawn for stroke joins respecting miter limit
References: 2d.shadow.render
Actual output:
Expected output:

Passed
